+++ apps/api/src/openapi/paths/models.ts
@@ -47,9 +47,11 @@ export const modelsPaths = {
                     id: "gpt-4o",
                     label: "GPT-4o",
                     providerId: "openai",
+                    providerName: "OpenAI",
                     apiShape: "openai-responses",
                     baseUrl: "https://api.openai.com/v1",
                     modelId: "gpt-4o",
+                    iconUrl: "openai",
                     source: "built-in",
                     enabled: true,
                     is_default: false,
