+++ apps/web/src/api/schema.d.ts
@@ -12563,9 +12563,11 @@ export interface operations {
                      *           "id": "gpt-4o",
                      *           "label": "GPT-4o",
                      *           "providerId": "openai",
+                     *           "providerName": "OpenAI",
                      *           "apiShape": "openai-responses",
                      *           "baseUrl": "https://api.openai.com/v1",
                      *           "modelId": "gpt-4o",
+                     *           "iconUrl": "openai",
                      *           "source": "built-in",
                      *           "enabled": true,
                      *           "is_default": false,
